feat(connection): move Local Gateway uninstall card to Connection page
The Local Gateway uninstall flow (and its MSIX warning, progress UI, and
result InfoBar) moves from Settings to a new card on the Connection page.
Connection is where it conceptually belongs — it operates on the gateway
runtime's lifecycle.

UI changes:
- New 'Local gateway on this PC' card on ConnectionPage, shown only when a
  local gateway is installed.
- Settings → Local Gateway expander deleted.
- Confirmation dialog text updated to be runtime-neutral ('local gateway
  runtime') so it reads correctly when future runtime options (Windows-
  native, etc.) ship alongside WSL.

Code:
- OnRemoveGateway / ShowUninstallError / ApplyUninstallUiState / state
  fields move from SettingsPage.xaml.cs to ConnectionPage.xaml.cs.
- SettingsPage shrinks; no longer references LocalGatewaySetup.
