+++ src/OpenClaw.Tray.WinUI/Pages/SettingsPage.xaml.cs
@@ -4,12 +4,8 @@ using Microsoft.UI.Xaml.Controls;
 using OpenClaw.Shared;
 using OpenClawTray.Helpers;
 using OpenClawTray.Services;
-using OpenClawTray.Services.LocalGatewaySetup;
 using OpenClawTray.Windows;
 using System;
-using System.IO;
-using System.Threading;
-using System.Threading.Tasks;
 
 namespace OpenClawTray.Pages;
 
@@ -19,15 +15,6 @@ public sealed partial class SettingsPage : Page
     private bool _initialized;
     private bool _saving;
     private bool _isDirty;
-    private bool _localGatewayInstalled;
-    private bool _uninstallInitiatedThisSession;
-    private CancellationTokenSource? _uninstallCts;
-
-    private enum UninstallUiState { Idle, InProgress, Success, Failure }
-
-    private const string GatewayIdleBodyText =
-        "Removes the WSL distro (OpenClawGateway), its disk image, autostart entry, and clears gateway credentials. Your MCP token is preserved. Onboarding will reset.";
-
 
     public SettingsPage()
     {
@@ -99,36 +86,7 @@ public sealed partial class SettingsPage : Page
         NotifyInfoCb.IsChecked = settings.NotifyInfo;
 
         // Pre-approve camera/screen toggles moved to Capabilities → row "Always allow".
-        LoadGatewaySection(settings);
     }
-
-    private void LoadGatewaySection(SettingsManager settings)
-    {
-        var localDataRoot = Path.Combine(
-            Environment.GetEnvironmentVariable("OPENCLAW_TRAY_LOCALAPPDATA_DIR")
-                ?? Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData),
-            "OpenClawTray");
-        var setupStatePath = Path.Combine(localDataRoot, "setup-state.json");
-
-        _localGatewayInstalled = File.Exists(setupStatePath)
-            || (settings.GatewayUrl?.StartsWith("ws://localhost", StringComparison.OrdinalIgnoreCase) == true);
-
-        LocalGatewayExpander.Visibility = ComputeLocalGatewaySectionVisibility();
-
-        // MSIX warning: Path A (conservative) — show when packaged AND gateway installed.
-        // TODO(commit-5): soften copy if Bostick's MSIX test confirms Path B (package-virtualized APPDATA).
-        MsixWarningBar.IsOpen = PackageHelper.IsPackaged && _localGatewayInstalled;
-    }
-
-    /// <summary>
-    /// Returns Visible when a local gateway exists OR an uninstall has been initiated this
-    /// view session (latch). The latch prevents the section from collapsing mid-flight when
-    /// the engine deletes setup-state.json before the result InfoBar is shown.
-    /// Resets on page navigation — section hides again on clean Settings re-open.
-    /// </summary>
-    private Visibility ComputeLocalGatewaySectionVisibility() =>
-        (_localGatewayInstalled || _uninstallInitiatedThisSession)
-            ? Visibility.Visible : Visibility.Collapsed;
 
     private void OnSave(object sender, RoutedEventArgs e)
     {
@@ -190,173 +148,6 @@ public sealed partial class SettingsPage : Page
         catch { }
     }
 
-    private async void OnRemoveGateway(object sender, RoutedEventArgs e)
-    {
-        // Build confirmation dialog content
-        var dialogContent = new StackPanel { Spacing = 8 };
-        dialogContent.Children.Add(new TextBlock
-        {
-            Text = "This will permanently remove the following:",
-            TextWrapping = TextWrapping.Wrap
-        });
-        dialogContent.Children.Add(new TextBlock
-        {
-            Text = "• WSL distro: OpenClawGateway (and its disk image)\n" +
-                   "• Autostart registry entry\n" +
-                   "• Gateway credentials (token and bootstrap token cleared)\n" +
-                   "• Setup state (onboarding will reset)",
-            TextWrapping = TextWrapping.Wrap,
-            Opacity = 0.7
-        });
-        dialogContent.Children.Add(new TextBlock
-        {
-            Text = "Preserved: Your MCP token and root device key are NOT deleted.\n" +
-                   "Removed: Local gateway identity credentials and registry records.",
-            TextWrapping = TextWrapping.Wrap,
-            Opacity = 0.7
-        });
-        dialogContent.Children.Add(new TextBlock
-        {
-            Text = "This cannot be undone.",
-            TextWrapping = TextWrapping.Wrap,
-            Margin = new Microsoft.UI.Xaml.Thickness(0, 4, 0, 0),
-            Opacity = 0.7
-        });
-
-        var dialog = new ContentDialog
-        {
-            Title = "Remove Local Gateway?",
-            Content = dialogContent,
-            PrimaryButtonText = "Remove Local Gateway",
-            CloseButtonText = "Cancel",
-            DefaultButton = ContentDialogButton.Close,
-            XamlRoot = XamlRoot
-        };
-
-        var dialogResult = await dialog.ShowAsync();
-        if (dialogResult != ContentDialogResult.Primary) return;
-
-        // Latch: keeps the section visible even after setup-state.json is deleted by the engine.
-        _uninstallInitiatedThisSession = true;
-        LocalGatewayExpander.Visibility = ComputeLocalGatewaySectionVisibility();
-
-        // --- Begin uninstall ---
-        ApplyUninstallUiState(UninstallUiState.InProgress);
-        UninstallResultBar.IsOpen = false;
-
-        _uninstallCts = new CancellationTokenSource();
-        try
-        {
-            if (_hub?.Settings == null)
-                throw new InvalidOperationException("Settings not available.");
-
-            var engine = LocalGatewayUninstall.Build(_hub.Settings, registry: _hub.GatewayRegistry);
-            var uninstallResult = await engine.RunAsync(
-                new LocalGatewayUninstallOptions
-                {
-                    DryRun = false,
-                    ConfirmDestructive = true
-                },
-                _uninstallCts.Token);
-
-            if (uninstallResult.Success)
-            {
-                ApplyUninstallUiState(UninstallUiState.Success);
-                UninstallResultBar.Severity = InfoBarSeverity.Success;
-                UninstallResultBar.Title = "Local gateway removed";
-                UninstallResultBar.Message = "Setup is reset; you can re-run the wizard from Onboarding.";
-                UninstallResultBar.ActionButton = null;
-                UninstallResultBar.IsOpen = true;
-            }
-            else
-            {
-                ApplyUninstallUiState(UninstallUiState.Failure);
-                var errorSummary = uninstallResult.Errors.Count > 0
-                    ? string.Join("; ", uninstallResult.Errors)
-                    : "Removal completed with unknown errors. Check logs for details.";
-                ShowUninstallError(errorSummary);
-            }
-        }
-        catch (OperationCanceledException)
-        {
-            ApplyUninstallUiState(UninstallUiState.Failure);
-            UninstallResultBar.Severity = InfoBarSeverity.Warning;
-            UninstallResultBar.Title = "Removal cancelled";
-            UninstallResultBar.Message = "Gateway may be in a partially-removed state. Review logs or retry.";
-            UninstallResultBar.ActionButton = null;
-            UninstallResultBar.IsOpen = true;
-        }
-        catch (Exception ex)
-        {
-            ApplyUninstallUiState(UninstallUiState.Failure);
-            ShowUninstallError(ex.Message);
-        }
-        finally
-        {
-            _uninstallCts?.Dispose();
-            _uninstallCts = null;
-        }
-    }
-
-    private void ShowUninstallError(string message)
-    {
-        var logsPath = Path.Combine(
-            Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData),
-            "OpenClawTray", "Logs");
-
-        var viewLogsButton = new Button { Content = "View Logs" };
-        viewLogsButton.Click += (_, _) =>
-        {
-            try { System.Diagnostics.Process.Start("explorer.exe", logsPath); } catch { }
-        };
-
-        UninstallResultBar.Severity = InfoBarSeverity.Error;
-        UninstallResultBar.Title = "Removal failed";
-        UninstallResultBar.Message = message;
-        UninstallResultBar.ActionButton = viewLogsButton;
-        UninstallResultBar.IsOpen = true;
-    }
-
-    private void ApplyUninstallUiState(UninstallUiState state)
-    {
-        switch (state)
-        {
-            case UninstallUiState.Idle:
-            case UninstallUiState.Failure:
-                RemoveGatewayButton.Content = "Remove Local Gateway";
-                RemoveGatewayButton.IsEnabled = true;
-                RemoveGatewayButton.Visibility = Visibility.Visible;
-                GatewayBodyText.Text = GatewayIdleBodyText;
-                break;
-
-            case UninstallUiState.InProgress:
-            {
-                var sp = new StackPanel { Orientation = Orientation.Horizontal, Spacing = 8 };
-                sp.Children.Add(new ProgressRing
-                {
-                    IsActive = true,
-                    Width = 16,
-                    Height = 16,
-                    VerticalAlignment = VerticalAlignment.Center
-                });
-                sp.Children.Add(new TextBlock
-                {
-                    Text = "Removing distro\u2026",
-                    VerticalAlignment = VerticalAlignment.Center
-                });
-                RemoveGatewayButton.Content = sp;
-                RemoveGatewayButton.IsEnabled = false;
-                RemoveGatewayButton.Visibility = Visibility.Visible;
-                GatewayBodyText.Text = "Removing the local gateway. This may take 10\u201330 seconds\u2026";
-                break;
-            }
-
-            case UninstallUiState.Success:
-                // Button hidden — success InfoBar carries the message.
-                // MSIX warning is moot once the gateway is gone.
-                RemoveGatewayButton.Visibility = Visibility.Collapsed;
-                MsixWarningBar.IsOpen = false;
-                break;
-        }
-    }
+    // OnRemoveGateway / ShowUninstallError / ApplyUninstallUiState / UninstallUiState
+    // moved to ConnectionPage along with the Local Gateway lifecycle UI.
 }
